adjust code pattern
