+++ components/survey_monkey/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@pipedream/survey_monkey",
+  "version": "0.0.1",
+  "description": "Pipedream SurveyMonkey Components",
+  "main": "survey_monkey.app.js",
+  "keywords": [
+    "pipedream",
+    "survey_monkey"
+  ],
+  "homepage": "https://pipedream.com/apps/survey-monkey",
+  "author": "Pipedream <support@pipedream.com> (https://pipedream.com/)",
+  "license": "MIT",
+  "dependencies": {
+    "uuid": "^8.3.2",
+    "@pipedream/platform": "^0.9.0"
+  }
+}
